requirements.txt
python-telegram-bot~=21.9
pillow~=11.0.0
pymongo~=4.10.1
python-dotenv~=1.0.1
python-daemon~=3.1.2